booking-service/requirements.txt
django
djangorestframework
django-prometheus
requests
